backend/auth-service/requirements.txt
Flask==2.3.2
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0
